+++ chain-signatures/node/Cargo.toml
@@ -39,7 +39,7 @@ opentelemetry_sdk = { version = "0.29.0", default-features = false, features = [
 opentelemetry-appender-tracing = "0.29.1"
 alloy-signer-local = "1.0.38"
 alloy-sol-types = "1.4.1"
-helios = { git = "https://github.com/a16z/helios", rev = "0a96f31125b7948bbe5c7ab73aa5699efae24902", optional = true }
+helios = { git = "https://github.com/a16z/helios", rev = "0a96f31125b7948bbe5c7ab73aa5699efae24902" }
 
 
 # workspace dependencies
@@ -98,8 +98,6 @@ test-log = { version = "0.2.12", features = ["log", "trace"] }
 
 [features]
 default = []
-# utilize ethereum light client via helios
-light_client = ["helios"]
 # utilized for benchmarking the node:
 bench = []
 # utilized by "integration-tests" package:
