+++ chain-signatures/node/src/indexer_eth/indexer_eth_helios.rs
@@ -1,0 +1,781 @@
+use alloy::eips::{BlockId, BlockNumberOrTag};
+
+use helios::common::types::{SubscriptionEvent, SubscriptionType};
+
+use helios::ethereum::{config::networks::Network, EthereumClient, EthereumClientBuilder};
+
+use std::path::PathBuf;
+
+use crate::indexer_eth::BlockAndRequests;
+use crate::indexer_eth::BlockToProcess;
+use crate::indexer_eth::EthConfig;
+use crate::sign_bidirectional::BidirectionalTxStatus;
+use crate::storage::app_data_storage::AppDataStorage;
+use alloy_sol_types::SolEvent;
+use near_account_id::AccountId;
+use std::collections::HashMap;
+use std::sync::Arc;
+use std::{str::FromStr, time::Instant};
+use tokio::sync::broadcast::error::RecvError;
+use tokio::sync::mpsc;
+use tokio::sync::RwLock;
+use tokio::time::Duration;
+
+use crate::indexer_eth::SignatureRequested;
+
+use crate::indexer_eth::parse_filtered_logs;
+use crate::indexer_eth::send_indexed_requests;
+use crate::protocol::{Chain, IndexedSignRequest};
+use crate::sign_bidirectional::BidirectionalTx;
+use crate::sign_bidirectional::BidirectionalTxId;
+use alloy::consensus::BlockHeader;
+use alloy::primitives::Address;
+use alloy::rpc::types::Log;
+
+// This is the maximum number of blocks that Helios can look back to
+const MAX_CATCHUP_BLOCKS: u64 = 8191;
+
+type BlockNumber = u64;
+
+const MAX_BLOCKS_TO_PROCESS: usize = 10000;
+
+fn blocks_to_process_channel() -> (mpsc::Sender<BlockToProcess>, mpsc::Receiver<BlockToProcess>) {
+    mpsc::channel(MAX_BLOCKS_TO_PROCESS)
+}
+
+const MAX_INDEXED_REQUESTS: usize = 1024;
+
+fn indexed_channel() -> (
+    mpsc::Sender<BlockAndRequests>,
+    mpsc::Receiver<BlockAndRequests>,
+) {
+    mpsc::channel(MAX_INDEXED_REQUESTS)
+}
+
+type BlockNumberAndHash = (u64, alloy::primitives::B256);
+
+const MAX_FAILED_BLOCKS: usize = 1024;
+
+fn failed_blocks_channel() -> (
+    mpsc::Sender<BlockNumberAndHash>,
+    mpsc::Receiver<BlockNumberAndHash>,
+) {
+    mpsc::channel(MAX_FAILED_BLOCKS)
+}
+
+const MAX_FINALIZED_BLOCKS: usize = 1024;
+
+fn finalized_block_channel() -> (mpsc::Sender<BlockNumber>, mpsc::Receiver<BlockNumber>) {
+    mpsc::channel(MAX_FINALIZED_BLOCKS)
+}
+
+pub async fn run(
+    eth: Option<EthConfig>,
+    sign_tx: mpsc::Sender<IndexedSignRequest>,
+    app_data_storage: AppDataStorage,
+    node_near_account_id: AccountId,
+    bidirectional_tx_map: Arc<RwLock<HashMap<BidirectionalTxId, BidirectionalTx>>>,
+) {
+    let Some(eth) = eth else {
+        tracing::warn!("ethereum indexer is disabled");
+        return;
+    };
+
+    let last_processed_block = app_data_storage
+        .last_processed_block_eth()
+        .await
+        .unwrap_or_else(|err| {
+            tracing::warn!("Failed to get last processed block: {err:?}");
+            None
+        });
+
+    let Ok(network) = Network::from_str(eth.network.as_str()) else {
+        tracing::error!("Network input incorrect: {}", eth.network);
+        return;
+    };
+
+    let client: EthereumClient = {
+        let builder = match EthereumClientBuilder::new()
+            .network(network)
+            .consensus_rpc(&eth.consensus_rpc_http_url)
+        {
+            Ok(builder) => builder,
+            Err(err) => {
+                tracing::error!("Failed to build consensus RPC: {err:?}");
+                return;
+            }
+        };
+
+        let builder = match builder.execution_rpc(&eth.execution_rpc_http_url) {
+            Ok(builder) => builder,
+            Err(err) => {
+                tracing::error!("Failed to build execution RPC: {err:?}");
+                return;
+            }
+        };
+
+        match builder
+            .data_dir(PathBuf::from(&eth.helios_data_path))
+            .with_file_db()
+            .build()
+        {
+            Ok(client) => client,
+            Err(err) => {
+                tracing::error!("Failed to build Helios client: {err:?}");
+                return;
+            }
+        }
+    };
+
+    tracing::info!("Built Helios client on network {}", network);
+
+    client.wait_synced().await;
+
+    tracing::info!("running ethereum indexer");
+
+    let mut block_heads_rx = match client.subscribe(SubscriptionType::NewHeads).await {
+        Ok(block_heads_rx) => block_heads_rx,
+        Err(err) => {
+            tracing::error!("Failed to subscribe to new block heads: {err:?}");
+            return;
+        }
+    };
+
+    let Ok(eth_contract_addr) = Address::from_str(&format!("0x{}", eth.contract_address)) else {
+        tracing::error!("Failed to parse contract address: {}", eth.contract_address);
+        return;
+    };
+    let total_timeout = Duration::from_secs(eth.total_timeout);
+
+    let (blocks_failed_send, blocks_failed_recv) = failed_blocks_channel();
+
+    let (requests_indexed_send, requests_indexed_recv) = indexed_channel();
+
+    let (finalized_block_send, finalized_block_recv) = finalized_block_channel();
+
+    let (blocks_to_process_send, mut blocks_to_process_recv) = blocks_to_process_channel();
+
+    let client = Arc::new(client);
+
+    let client_clone = Arc::clone(&client);
+    tokio::spawn(async move {
+        tracing::info!("Spawned task to refresh the latest finalized block");
+        refresh_finalized_block(
+            &client_clone,
+            finalized_block_send.clone(),
+            eth.refresh_finalized_interval,
+        )
+        .await;
+    });
+
+    let near_account_id_clone = node_near_account_id.clone();
+    let client_clone = Arc::clone(&client);
+    tokio::spawn(async move {
+        tracing::info!("Spawned task to send indexed requests to send queue");
+        send_requests_when_final(
+            &client_clone,
+            requests_indexed_recv,
+            finalized_block_recv,
+            sign_tx.clone(),
+            app_data_storage.clone(),
+            near_account_id_clone.clone(),
+        )
+        .await;
+    });
+
+    let near_account_id_clone = node_near_account_id.clone();
+    let requests_indexed_send_clone = requests_indexed_send.clone();
+    let blocks_failed_send_clone = blocks_failed_send.clone();
+    let client_clone = Arc::clone(&client);
+    let bidirectional_tx_map_clone = bidirectional_tx_map.clone();
+    tokio::spawn(async move {
+        tracing::info!("Spawned task to retry failed blocks");
+        retry_failed_blocks(
+            blocks_failed_recv,
+            blocks_failed_send_clone,
+            &client_clone,
+            eth_contract_addr,
+            near_account_id_clone,
+            requests_indexed_send_clone,
+            total_timeout,
+            bidirectional_tx_map_clone,
+        )
+        .await;
+    });
+
+    let blocks_to_process_send_clone = blocks_to_process_send.clone();
+    if let Some(last_processed_block) = last_processed_block {
+        let Ok(SubscriptionEvent::NewHeads(latest_block)) = block_heads_rx.recv().await else {
+            tracing::warn!("Failed to receive latest block head");
+            return;
+        };
+        let end_block_number = latest_block.header.number;
+        add_catchup_blocks_to_process(
+            blocks_to_process_send_clone,
+            last_processed_block,
+            end_block_number,
+        )
+        .await;
+    }
+
+    let blocks_to_process_send_clone = blocks_to_process_send.clone();
+    tokio::spawn(async move {
+        tracing::info!("Spawned task to add new blocks to process");
+        add_new_block_to_process(block_heads_rx, blocks_to_process_send_clone).await;
+    });
+
+    let mut interval = tokio::time::interval(Duration::from_millis(200));
+    let requests_indexed_send_clone = requests_indexed_send.clone();
+    loop {
+        let Some(block_to_process) = blocks_to_process_recv.recv().await else {
+            interval.tick().await;
+            continue;
+        };
+        let (block_number, block_hash, is_catchup) = match block_to_process {
+            BlockToProcess::Catchup(block_number) => {
+                let block = fetch_block(
+                    &client,
+                    BlockId::Number(BlockNumberOrTag::Number(block_number)),
+                    5,
+                    Duration::from_millis(200),
+                )
+                .await;
+                if let Some(block) = block {
+                    (block.header.number, block.header.hash, true)
+                } else {
+                    continue;
+                }
+            }
+            BlockToProcess::NewBlock((block_number, block_hash)) => {
+                (block_number, block_hash, false)
+            }
+        };
+        let bidirectional_tx_map_clone = bidirectional_tx_map.clone();
+        if let Err(err) = process_block(
+            block_number,
+            block_hash,
+            &client,
+            eth_contract_addr,
+            node_near_account_id.clone(),
+            requests_indexed_send_clone.clone(),
+            total_timeout,
+            bidirectional_tx_map_clone,
+        )
+        .await
+        {
+            tracing::warn!("Eth indexer failed to process block number {block_number}: {err:?}");
+            add_failed_block(blocks_failed_send.clone(), block_number, block_hash).await;
+            continue;
+        }
+        if block_number % 10 == 0 {
+            if is_catchup {
+                tracing::info!("Processed catchup block number {block_number}");
+            } else {
+                tracing::info!("Processed new block number {block_number}");
+            }
+        }
+        crate::metrics::LATEST_BLOCK_NUMBER
+            .with_label_values(&[Chain::Ethereum.as_str(), node_near_account_id.as_str()])
+            .set(block_number as i64);
+    }
+}
+
+#[allow(clippy::too_many_arguments)]
+async fn retry_failed_blocks(
+    mut blocks_failed_rx: mpsc::Receiver<BlockNumberAndHash>,
+    blocks_failed_tx: mpsc::Sender<BlockNumberAndHash>,
+    client: &Arc<EthereumClient>,
+    eth_contract_addr: Address,
+    node_near_account_id: AccountId,
+    requests_indexed: mpsc::Sender<BlockAndRequests>,
+    total_timeout: Duration,
+    bidirectional_tx_map: Arc<RwLock<HashMap<BidirectionalTxId, BidirectionalTx>>>,
+) {
+    loop {
+        let Some((block_number, block_hash)) = blocks_failed_rx.recv().await else {
+            tracing::warn!("Failed to receive block and requests from requests_indexed");
+            break;
+        };
+        if let Err(err) = process_block(
+            block_number,
+            block_hash,
+            client,
+            eth_contract_addr,
+            node_near_account_id.clone(),
+            requests_indexed.clone(),
+            total_timeout,
+            bidirectional_tx_map.clone(),
+        )
+        .await
+        {
+            tracing::warn!("Retry failed for block {block_number}: {err:?}");
+            add_failed_block(blocks_failed_tx.clone(), block_number, block_hash).await;
+        } else {
+            tracing::info!("Successfully retried block: {block_number}");
+        }
+    }
+}
+
+async fn add_failed_block(
+    blocks_failed: mpsc::Sender<BlockNumberAndHash>,
+    block_number: u64,
+    block_hash: alloy::primitives::B256,
+) {
+    blocks_failed
+        .send((block_number, block_hash))
+        .await
+        .unwrap_or_else(|err| {
+            tracing::warn!("Failed to send failed block: {:?}", err);
+        });
+}
+
+async fn add_new_block_to_process(
+    mut block_heads_rx: tokio::sync::broadcast::Receiver<
+        SubscriptionEvent<helios::ethereum::spec::Ethereum>,
+    >,
+    blocks_to_process: mpsc::Sender<BlockToProcess>,
+) {
+    let mut interval = tokio::time::interval(Duration::from_millis(200));
+    let mut receiver_state_update_timestamp = Instant::now();
+    loop {
+        interval.tick().await;
+        if block_heads_rx.is_empty()
+            && receiver_state_update_timestamp.elapsed() > Duration::from_secs(60)
+        {
+            tracing::warn!("No new block heads received for 60 seconds, waiting...");
+            receiver_state_update_timestamp = Instant::now();
+        }
+        let new_block_head = match block_heads_rx.recv().await {
+            Ok(new_block_head) => new_block_head,
+            Err(RecvError::Lagged(lagged_count)) => {
+                tracing::warn!(
+                    "Eth indexer failed to receive latest block header: block heads stream lagged too far behind, lagged count: {lagged_count}"
+                );
+                continue;
+            }
+            Err(RecvError::Closed) => {
+                tracing::error!(
+                    "Eth indexer failed to receive latest block header: block heads stream closed"
+                );
+                // TODO: add a retry mechanism for closed block heads stream
+                break;
+            }
+        };
+        receiver_state_update_timestamp = Instant::now();
+        let SubscriptionEvent::NewHeads(new_block) = new_block_head;
+        let block_number = new_block.header.number;
+        let block_hash = new_block.header.hash;
+        if block_number % 10 == 0 {
+            tracing::info!("Received new block head: {block_number}");
+        }
+        if let Err(err) = blocks_to_process
+            .send(BlockToProcess::NewBlock((block_number, block_hash)))
+            .await
+        {
+            tracing::warn!("Failed to send block to process: {err:?}");
+        }
+    }
+}
+
+async fn add_catchup_blocks_to_process(
+    blocks_to_process: mpsc::Sender<BlockToProcess>,
+    start_block_number: u64,
+    end_block_number: u64,
+) {
+    // helios can only go back maximum MAX_CATCHUP_BLOCKS blocks, so we need to adjust the start block number if it's too far behind
+    let helios_oldest_block_number = end_block_number.saturating_sub(MAX_CATCHUP_BLOCKS);
+    let start_block_number = if start_block_number < helios_oldest_block_number {
+        tracing::warn!(
+            "Start block number {start_block_number} is too far behind the latest block {end_block_number}, adjusting to {helios_oldest_block_number}"
+        );
+        helios_oldest_block_number
+    } else {
+        start_block_number
+    };
+
+    for block_number in start_block_number..=end_block_number {
+        if let Err(err) = blocks_to_process
+            .send(BlockToProcess::Catchup(block_number))
+            .await
+        {
+            tracing::warn!("Failed to send block to process: {err:?}");
+        }
+    }
+}
+
+// retry getting block from helios with exponential backoff
+
+async fn fetch_block(
+    helios_client: &Arc<EthereumClient>,
+    block_id: BlockId,
+    max_retries: u8,
+    base_delay: Duration,
+) -> Option<alloy::rpc::types::Block> {
+    let mut retries = 0;
+    loop {
+        match helios_client.get_block(block_id, false).await {
+            Ok(Some(block)) => return Some(block),
+            Ok(None) => {
+                tracing::warn!("Block {block_id} not found from Helios client");
+                return None;
+            }
+            Err(e) => {
+                if retries < max_retries {
+                    retries += 1;
+                    let delay = base_delay * 2u32.pow((retries - 1) as u32);
+                    tracing::warn!(
+                        "Failed to fetch block number {block_id} from Helios client: {:?}, retrying",
+                        e
+                    );
+                    tokio::time::sleep(delay).await;
+                    continue;
+                }
+                tracing::warn!(
+                    "Failed to fetch block number {block_id} from Helios client: {:?}, exceeded maximum retry",
+                    e
+                );
+                return None;
+            }
+        }
+    }
+}
+
+/// Polls for the latest finalized block and update finalized block channel.
+async fn refresh_finalized_block(
+    helios_client: &Arc<EthereumClient>,
+    finalized_block_send: mpsc::Sender<BlockNumber>,
+    refresh_finalized_interval: u64,
+) {
+    let mut interval = tokio::time::interval(Duration::from_millis(refresh_finalized_interval));
+    let mut final_block_number: Option<BlockNumber> = None;
+
+    loop {
+        interval.tick().await;
+        tracing::info!("Refreshing finalized epoch");
+
+        let new_finalized_block = match fetch_block(
+            helios_client,
+            BlockId::Number(BlockNumberOrTag::Finalized),
+            5,
+            Duration::from_millis(200),
+        )
+        .await
+        {
+            Some(block) => block,
+            None => {
+                continue;
+            }
+        };
+
+        let new_final_block_number = new_finalized_block.header.number;
+        tracing::info!(
+            "New finalized block number: {new_final_block_number}, last finalized block number: {final_block_number:?}"
+        );
+
+        if final_block_number.is_none_or(|n| new_final_block_number > n) {
+            tracing::info!("Found new finalized block!");
+            if let Err(err) = finalized_block_send.send(new_final_block_number).await {
+                tracing::warn!("Failed to send finalized block: {err:?}");
+                continue;
+            }
+            final_block_number.replace(new_final_block_number);
+            continue;
+        }
+
+        let Some(last_final_block_number) = final_block_number else {
+            continue;
+        };
+
+        if new_final_block_number < last_final_block_number {
+            tracing::warn!(
+                "New finalized block number overflowed range of u64 and has wrapped around!"
+            );
+        }
+
+        if last_final_block_number == new_final_block_number {
+            tracing::info!("No new finalized block");
+        }
+    }
+}
+
+#[allow(clippy::too_many_arguments)]
+async fn process_block(
+    block_number: u64,
+    block_hash: alloy::primitives::B256,
+    client: &Arc<EthereumClient>,
+    eth_contract_addr: Address,
+    node_near_account_id: AccountId,
+    requests_indexed: mpsc::Sender<BlockAndRequests>,
+    total_timeout: Duration,
+    bidirectional_tx_map: Arc<RwLock<HashMap<BidirectionalTxId, BidirectionalTx>>>,
+) -> anyhow::Result<()> {
+    tracing::info!(
+        "Processing block number {} with hash {:?}",
+        block_number,
+        block_hash
+    );
+    let start = Instant::now();
+    let block_receipts_result = client
+        .get_block_receipts(BlockId::Number(BlockNumberOrTag::Number(block_number)))
+        .await;
+    crate::metrics::ETH_BLOCK_RECEIPT_LATENCY
+        .with_label_values(&[node_near_account_id.as_str()])
+        .observe(start.elapsed().as_millis() as f64);
+    let Some(block_receipts) = block_receipts_result.map_err(|err| {
+        anyhow::anyhow!(
+            "Failed to get block receipts for block number {block_number}: {:?}",
+            err
+        )
+    })?
+    else {
+        tracing::info!("no receipts for block number {block_number}");
+        return Ok(());
+    };
+
+    let pending_txs: HashMap<BidirectionalTxId, BidirectionalTx> = {
+        bidirectional_tx_map
+            .read()
+            .await
+            .iter()
+            .filter(|(_, tx)| tx.status == BidirectionalTxStatus::Pending)
+            .map(|(id, tx)| (*id, tx.clone()))
+            .collect()
+    };
+    let mut respond_bidirectional_requests: Vec<IndexedSignRequest> = Vec::new();
+    for receipt in &block_receipts {
+        let Some(pending_tx) = pending_txs.get(&receipt.transaction_hash.into()) else {
+            continue;
+        };
+        let status = receipt.status();
+        tracing::info!(
+            "Tx {} found in block {block_number}: {}",
+            receipt.transaction_hash,
+            if status { "success" } else { "failed" }
+        );
+        let client_clone = client.clone();
+        let mut pending_tx = pending_tx.clone();
+        pending_tx.status = if status {
+            BidirectionalTxStatus::Success
+        } else {
+            BidirectionalTxStatus::Failed
+        };
+
+        let pending_tx_clone = pending_tx.clone();
+        let completed_tx =
+            crate::respond_bidirectional::CompletedTx::new(pending_tx_clone, block_number);
+
+        let respond_bidirectional_sign_request: Option<IndexedSignRequest> = completed_tx
+            .create_sign_request_from_completed_tx(&client_clone, 6, total_timeout)
+            .await;
+        if let Some(respond_bidirectional_sign_request) = respond_bidirectional_sign_request {
+            respond_bidirectional_requests.push(respond_bidirectional_sign_request);
+            bidirectional_tx_map
+                .write()
+                .await
+                .insert(receipt.transaction_hash.into(), pending_tx);
+        } else {
+            // failed to create sign request from completed tx, remove the tx from the map
+            // TODO: we need to implement a better solution for not finding such txs in helios. possible: https://github.com/sig-net/mpc/issues/499
+            tracing::warn!(
+                "Failed to create sign request from completed tx, removing tx from map: {:?}",
+                receipt.transaction_hash
+            );
+            bidirectional_tx_map
+                .write()
+                .await
+                .remove(&receipt.transaction_hash.into());
+        }
+    }
+
+    let remaining_pending_txs: HashMap<BidirectionalTxId, BidirectionalTx> = {
+        bidirectional_tx_map
+            .read()
+            .await
+            .iter()
+            .filter(|(_, tx)| tx.status == BidirectionalTxStatus::Pending)
+            .map(|(id, tx)| (*id, tx.clone()))
+            .collect()
+    };
+
+    for (tx_id, mut tx) in remaining_pending_txs {
+        let current_nonce = client
+            .get_nonce(
+                tx.from_address,
+                BlockId::Number(BlockNumberOrTag::Number(block_number)),
+            )
+            .await;
+        if current_nonce.is_err() {
+            tracing::warn!("Failed to get current nonce: {:?}", current_nonce.err());
+            continue;
+        }
+        let current_nonce = current_nonce.unwrap();
+        if tx.nonce < current_nonce {
+            tracing::warn!(
+                "Nonce too low for tx {:?}: expected {}, got {}",
+                tx_id,
+                tx.nonce,
+                current_nonce
+            );
+            tx.status = BidirectionalTxStatus::Failed;
+            let client_clone = client.clone();
+
+            let pending_tx = tx.clone();
+
+            let completed_tx =
+                crate::respond_bidirectional::CompletedTx::new(pending_tx, block_number);
+
+            let respond_bidirectional_sign_request: Option<IndexedSignRequest> = completed_tx
+                .create_sign_request_from_completed_tx(&client_clone, 6, total_timeout)
+                .await;
+            if let Some(respond_bidirectional_sign_request) = respond_bidirectional_sign_request {
+                respond_bidirectional_requests.push(respond_bidirectional_sign_request);
+                bidirectional_tx_map.write().await.insert(tx_id, tx);
+            } else {
+                // failed to create sign request from completed tx, remove the tx from the map
+                tracing::warn!(
+                    "Failed to create sign request from completed tx, removing tx from map: {:?}",
+                    tx_id
+                );
+                bidirectional_tx_map.write().await.remove(&tx_id);
+            }
+        }
+    }
+
+    let filtered_logs: Vec<Log> = block_receipts
+        .into_iter()
+        .filter_map(|receipt| receipt.as_ref().as_receipt().cloned())
+        .flat_map(|receipt| {
+            receipt.logs.into_iter().filter(|log| {
+                log.address() == eth_contract_addr
+                    && log
+                        .topic0()
+                        .is_some_and(|topic0| *topic0 == SignatureRequested::SIGNATURE_HASH)
+            })
+        })
+        .collect();
+
+    let mut all_sign_requests = Vec::new();
+    if !filtered_logs.is_empty() {
+        all_sign_requests.extend(parse_filtered_logs(filtered_logs, total_timeout));
+    }
+    if !respond_bidirectional_requests.is_empty() {
+        all_sign_requests.extend(respond_bidirectional_requests);
+    }
+
+    if all_sign_requests.is_empty() {
+        return Ok(());
+    }
+
+    requests_indexed
+        .send(BlockAndRequests::new(
+            block_number,
+            block_hash,
+            all_sign_requests.clone(),
+        ))
+        .await
+        .map_err(|err| anyhow::anyhow!("Failed to send indexed requests: {:?}", err))?;
+
+    let block_timestamp = client
+        .get_block(
+            BlockId::Number(BlockNumberOrTag::Number(block_number)),
+            false,
+        )
+        .await
+        .ok()
+        .and_then(|block| block.map(|b| b.header.timestamp()));
+
+    for request in &all_sign_requests {
+        if let Some(block_timestamp) = block_timestamp {
+            crate::metrics::INDEXER_DELAY
+                .with_label_values(&[Chain::Ethereum.as_str(), node_near_account_id.as_str()])
+                .observe(
+                    crate::util::duration_between_unix(
+                        block_timestamp,
+                        request.unix_timestamp_indexed,
+                    )
+                    .as_secs() as f64,
+                );
+        }
+    }
+
+    Ok(())
+}
+
+/// Sends a request to the sign queue when the block where the request is in is finalized.
+async fn send_requests_when_final(
+    helios_client: &Arc<EthereumClient>,
+    mut requests_indexed: mpsc::Receiver<BlockAndRequests>,
+    mut finalized_block_rx: mpsc::Receiver<BlockNumber>,
+    sign_tx: mpsc::Sender<IndexedSignRequest>,
+    app_data_storage: AppDataStorage,
+    node_near_account_id: AccountId,
+) {
+    let mut finalized_block_number: Option<BlockNumber> = None;
+    let mut last_processed_block: Option<BlockNumber> = app_data_storage
+        .last_processed_block_eth()
+        .await
+        .unwrap_or_else(|err| {
+            tracing::warn!("Failed to fetch last processed block: {err:?}, setting to None");
+            None
+        });
+
+    loop {
+        let Some(BlockAndRequests {
+            block_number,
+            block_hash,
+            indexed_requests,
+        }) = requests_indexed.recv().await
+        else {
+            tracing::error!("Failed to receive indexed requests");
+            return;
+        };
+
+        // Wait for finalized block if needed
+        while finalized_block_number.is_none_or(|n| block_number > n) {
+            let Some(new_finalized_block) = finalized_block_rx.recv().await else {
+                tracing::error!("Failed to receive finalized blocks");
+                return;
+            };
+            finalized_block_number.replace(new_finalized_block);
+        }
+
+        // Verify block hash and send requests
+        let block = fetch_block(
+            helios_client,
+            block_number.into(),
+            5,
+            Duration::from_millis(200),
+        )
+        .await;
+
+        let Some(block) = block else {
+            tracing::warn!("Block {block_number} not found from Helios client, skipping this block and its requests");
+            continue;
+        };
+
+        if block.header.hash == block_hash {
+            tracing::info!("Block {block_number} is finalized!");
+            send_indexed_requests(
+                indexed_requests,
+                sign_tx.clone(),
+                node_near_account_id.clone(),
+            );
+            if last_processed_block.is_none_or(|n| n < block_number) {
+                if let Err(err) = app_data_storage
+                    .set_last_processed_block_eth(block_number)
+                    .await
+                {
+                    tracing::warn!("Failed to set last processed block: {err:?}");
+                }
+                last_processed_block.replace(block_number);
+            }
+        } else {
+            // no special handling for chain reorg, just log the error
+            // This is because when such chain reorg happens, the new canonical chain will have already been emitted by helios's block header stream, and we can safely skip this block here.
+            tracing::error!(
+                "Block {block_number} hash mismatch: expected {block_hash:?}, got {:?}. Chain re-orged.",
+                block.header.hash
+            );
+        }
+    }
+}
