+++ chain-signatures/node/src/respond_bidirectional.rs
@@ -1,20 +1,20 @@
 use crate::protocol::{Chain, IndexedSignRequest};
 use crate::sign_bidirectional::BidirectionalTx;
 use crate::sign_bidirectional::BidirectionalTxId;
-#[cfg(feature = "light_client")]
+
 use crate::sign_bidirectional::BidirectionalTxStatus;
-#[cfg(feature = "light_client")]
+
 use crate::sign_bidirectional::TransactionOutput;
-#[cfg(feature = "light_client")]
+
 use alloy::consensus::Transaction;
-#[cfg(feature = "light_client")]
+
 use alloy::eips::{BlockId, BlockNumberOrTag};
-#[cfg(feature = "light_client")]
+
 use alloy::primitives::Address;
 use alloy::primitives::Bytes;
-#[cfg(feature = "light_client")]
+
 use alloy::rpc::types::TransactionRequest;
-#[cfg(feature = "light_client")]
+
 use helios::ethereum::EthereumClient;
 use k256::Scalar;
 use mpc_crypto::ScalarExt;
@@ -36,7 +36,7 @@ pub enum SerDeserFormat {
 
 pub struct CompletedTx {
     tx: BidirectionalTx,
-    #[cfg(feature = "light_client")]
+
     block_number: u64,
 }
 
@@ -49,16 +49,11 @@ pub struct RespondBidirectionalTx {
 pub type RespondBidirectionalSerializedOutput = Vec<u8>;
 
 impl CompletedTx {
-    #[cfg_attr(not(feature = "light_client"), allow(unused_variables))]
+    #[allow(unused_variables)]
     pub fn new(tx: BidirectionalTx, block_number: u64) -> Self {
-        Self {
-            tx,
-            #[cfg(feature = "light_client")]
-            block_number,
-        }
+        Self { tx, block_number }
     }
 
-    #[cfg(not(feature = "light_client"))]
     pub(crate) async fn create_failed_sign_request_without_light_client(
         &self,
         chain: Chain,
@@ -68,7 +63,6 @@ impl CompletedTx {
             .await
     }
 
-    #[cfg(not(feature = "light_client"))]
     pub(crate) fn create_sign_request_from_serialized_output(
         &self,
         chain: Chain,
@@ -82,7 +76,6 @@ impl CompletedTx {
         )
     }
 
-    #[cfg(feature = "light_client")]
     pub async fn create_sign_request_from_completed_tx(
         &self,
         helios_client: &Arc<EthereumClient>,
@@ -116,7 +109,6 @@ impl CompletedTx {
         }
     }
 
-    #[cfg(feature = "light_client")]
     async fn process_completed_tx(
         &self,
         helios_client: &Arc<EthereumClient>,
@@ -171,7 +163,6 @@ impl CompletedTx {
         Ok(sign_request)
     }
 
-    #[cfg(feature = "light_client")]
     async fn process_success_tx(
         &self,
         helios_client: &Arc<EthereumClient>,
@@ -248,7 +239,6 @@ impl CompletedTx {
         })
     }
 
-    #[cfg(feature = "light_client")]
     async fn extract_success_tx_output(
         &self,
         helios_client: &Arc<EthereumClient>,
@@ -295,7 +285,6 @@ fn calculate_respond_bidirectional_hash_message(
     alloy::primitives::keccak256(&combined).into()
 }
 
-#[cfg(feature = "light_client")]
 async fn fetch_call_result(
     helios_client: &Arc<EthereumClient>,
     from_address: Address,
@@ -331,7 +320,6 @@ async fn fetch_call_result(
     }
 }
 
-#[cfg(feature = "light_client")]
 async fn fetch_tx_from_helios(
     helios_client: &Arc<EthereumClient>,
     tx_id: BidirectionalTxId,
